Move GIT integration from bodiless-core to bodiless-git
