+++ packages/bodiless-git/README.md
@@ -1,0 +1,4 @@
+# `@bodiless/git`
+
+Default Git Integration for bodiless.
+
